Updated to zop 4.0 packet spec
Will need to be merged in conjunction with sst-forza-elements to keep them in sync.

Updated the test script to use larger flit sizes to match HW specs.

+++ src/ZOPNET.cc
@@ -295,7 +295,7 @@ void zopNIC::send_zone_barrier( unsigned Hart, unsigned endpoints ) {
       ev->setSrcPrec( (uint8_t) ( getPrecinctID() ) );
 
       std::vector<uint64_t> payload;
-      payload.push_back( (uint64_t) ( barrierSense[Hart] ) );
+      payload.push_back( barrierSense[Hart] );
       ev->setPayload( payload );
       ev->encodeEvent();
